corrected remarks

+++ project/src/matrix_transpose_multi_thread.c
@@ -137,7 +137,7 @@ matrix_t* transp(matrix_t* matrix) {
 
         exit(EXIT_SUCCESS);
     }
-    size_t status;
+    int status;
     for (size_t i = 0; i < process_count; i++) {
         if (waitpid(pids[i], &status, 0) != pids[i]) {
             free_matrix(new_matrix);
